+++ mon.h
@@ -534,7 +534,7 @@ void			 cocb_tourjob(int);
 void			 cam_move(int, float);
 void			 cam_revolve(struct fvec *, float, float);
 void			 cam_rotate(const struct fvec *, const struct fvec *, int, int);
-void			 cam_roll(float);
+void			 cam_roll(double);
 void			 cam_look(void);
 
 /* capture.c */
